Refactor ReleaseManager into a separate class per command
This is in preparation for creating a new "push" command. I couldn't
bear adding to the mess of a huge Program.cs
