+++ tools/Google.Cloud.Tools.ReleaseManager/ShowCommand.cs
@@ -1,0 +1,34 @@
+﻿// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+
+namespace Google.Cloud.Tools.ReleaseManager
+{
+    public sealed class ShowCommand : CommandBase
+    {
+        public ShowCommand()
+            : base("show", "Show the versions changes in the current PR")
+        {
+        }
+
+        protected override void ExecuteImpl(string[] args)
+        {
+            foreach (var change in FindChangedVersions())
+            {
+                Console.WriteLine(change);
+            }
+        }
+    }
+}
